pyproject.toml
# ⚠️ IMPORTANT
# This is proprietary code of Outhad AI (outhad.com) developed by Mohammad Tanzil Idrisi

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "unified_query"
version = "0.1.0"
description = "A middleware package for unified SQL, Cypher, and vector operations"
requires-python = ">=3.8"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "neo4j>=5.0.0",
    "faiss-cpu>=1.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "asyncio>=3.4.3",
    "typing-extensions>=4.0.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/unified_query"]